requirements.txt
# NeuroVision AI - Brain Tumor Detection
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer
vit-pytorch>=1.6.0

# Web application
streamlit>=1.32.0

# Image processing
Pillow>=9.0.0

# Data science & visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ML utilities
scikit-learn>=1.3.0

# Training utilities
tqdm>=4.65.0
prettytable>=3.8.0
arrow>=1.2.3

# HTTP requests (Ollama chatbot)
requests>=2.31.0
